Moved classes used explicitly for testing to 'tests' subdirectory
